split train_test_split module file into several modules
